requirements.txt
numpy>=1.16.1
pandas>=0.24.1
click>=7.0
h5py>=2.9.0
matplotlib>=3.0.2
multiprocess>=0.70.7
pysam>=0.15.0
pyyaml>=5.1
